requirements.txt
pathlib
pillow
pytesseract
edge-tts
